fix(StyledCheckbox): StyledCheckbox вынесен в компонент FormCheckbox
